Rearrange the library
This moves some types into more sensible modules and renames a few things whose
meaning has changed over time.

The user-observable change here is that `InstructionConstraints` and
`ArchConstraints` have switched meanings (they were backwards given their
contents/type variables).
