+++ renovate/renovate.cabal
@@ -8,7 +8,7 @@ synopsis:            An interface for binary analysis and rewriting
 license:             BSD3
 license-file:        LICENSE
 author:              Tristan Ravitch
-maintainer:          tristan@nochair.net
+maintainer:          tristan@galois.com
 -- copyright:           
 category:            Disassembler
 build-type:          Simple
@@ -21,24 +21,25 @@ library
                        Renovate.BinaryFormat.ELF.Common.Internal
   other-modules:
                        Renovate.Arch
-                       Renovate.BasicBlock.Assemble
-                       Renovate.BasicBlock.Types
+                       Renovate.Assemble
+                       Renovate.Analysis.FunctionRecovery
+                       Renovate.ABI
                        Renovate.Config
+                       Renovate.Core.Address
+                       Renovate.Core.BasicBlock
+                       Renovate.Core.Instruction
+                       Renovate.Core.Layout
+                       Renovate.Core.Relocation
                        Renovate.Diagnostic
                        Renovate.Metrics
+                       Renovate.Redirect
                        Renovate.Redirect.Concretize
                        Renovate.Redirect.Internal
                        Renovate.Redirect.LayoutBlocks
-                       Renovate.Redirect.LayoutBlocks.Compact
-                       Renovate.Redirect.LayoutBlocks.Types
                        Renovate.Redirect.Monad
                        Renovate.Redirect.ReifyFallthrough
                        Renovate.Redirect.Symbolize
                        Renovate.Rewrite
-                       Renovate.ABI
-                       Renovate.Address
-                       Renovate.Analysis.FunctionRecovery
-                       Renovate.BasicBlock
                        Renovate.BinaryFormat.ELF
                        Renovate.BinaryFormat.ELF.Common
                        Renovate.BinaryFormat.ELF.BSS
@@ -51,11 +52,10 @@ library
                        Renovate.Recovery
                        Renovate.Recovery.Overlap
                        Renovate.Recovery.SymbolMap
-                       Renovate.Redirect
   -- other-extensions:    
   build-depends:       base >=4.10 && <5,
                        panic >= 0.4 && < 0.5,
-                       bytestring,
+                       bytestring >= 0.9 && < 0.12,
                        binary,
                        text,
                        containers >= 0.5 && < 0.7,
@@ -66,9 +66,8 @@ library
                        mwc-random >= 0.13.6.0 && < 0.15,
                        exceptions >= 0.7 && < 0.11,
                        vector >= 0.10 && < 0.13,
-                       lens >= 4 && < 5,
+                       lens >= 4 && < 6,
                        generic-lens >= 0.5 && < 1.1,
-                       located-base >= 0.1 && < 0.2,
                        lumberjack >= 0.1 && < 1.1,
                        union-find >= 0.2 && < 0.3,
                        prettyprinter >= 1.7 && < 1.8,
@@ -91,9 +90,9 @@ executable run-discovery
   main-is:             run-discovery.hs
   default-language:    Haskell2010
   build-depends:       base >= 4.10 && < 5,
-                       bytestring >= 0.9 && < 0.11,
-                       containers >= 0.5 && < 0.7,
-                       lens >= 4 && < 5,
+                       bytestring,
+                       containers,
+                       lens,
                        elf-edit,
                        macaw-base,
                        macaw-loader,
